+++ example/lib/custom_marker.dart
@@ -1,6 +1,7 @@
 import 'dart:math';
 
 import 'package:flutter/material.dart';
+import 'package:vietmap_gl_example/constant.dart';
 
 import 'package:vietmap_flutter_gl/vietmap_flutter_gl.dart';
 
@@ -36,8 +37,7 @@ class CustomMarkerState extends State<CustomMarker> {
     return new Scaffold(
       body: Stack(children: [
         VietmapGL(
-          styleString:
-              "https://maps.vietmap.vn/api/maps/light/styles.json?apikey=YOUR_API_KEY_HERE",
+          styleString: YOUR_STYLE_URL_HERE,
           trackCameraPosition: true,
           onMapCreated: _onMapCreated,
           initialCameraPosition: const CameraPosition(