+++ vietmap_gl_platform_interface/lib/src/helper/vietmap_polyline_utils.dart
@@ -54,7 +54,7 @@ class NearestLatLngResult {
 }
 
 class VietmapPolyline {
-  static NearestLatLngResult _nearestLatLngOnLine(
+  static NearestLatLngResult? _nearestLatLngOnLine(
     List<LatLng> line,
     LatLng point, [
     Unit unit = Unit.kilometers,
@@ -88,23 +88,23 @@ class VietmapPolyline {
       );
 
       final heightDistance = max(start.distance, stop.distance);
-      final direction = bearing(startLatLng, stopLatLng);
+      final direction = _bearing(startLatLng, stopLatLng);
 
-      final perpendicular1 = destination(
+      final perpendicular1 = _destination(
         point,
         heightDistance,
         direction + 90,
         unit,
       );
 
-      final perpendicular2 = destination(
+      final perpendicular2 = _destination(
         point,
         heightDistance,
         direction - 90,
         unit,
       );
 
-      final intersectionLatLng = intersects(
+      final intersectionLatLng = _intersects(
         [perpendicular1, perpendicular2],
         [startLatLng, stopLatLng],
       );
@@ -139,10 +139,10 @@ class VietmapPolyline {
     /// A `List<LatLng>` is guaranteed to have at least two points and thus a
     /// nearest point has to exist.
 
-    return isGetStop ? stopP! : nearest!;
+    return isGetStop ? stopP : nearest;
   }
 
-  static NearestLatLngResult nearestLatLngOnLine(
+  static NearestLatLngResult? nearestLatLngOnLine(
     List<LatLng> line,
     LatLng point, [
     Unit unit = Unit.kilometers,
@@ -157,11 +157,11 @@ class VietmapPolyline {
     bool snapInputLatLngToResult = true,
   }) {
     var res = _nearestLatLngOnLine(line, point, unit, true);
-    var line1 = line.sublist(0, res.index + 1);
+    var line1 = line.sublist(0, res?.index ?? -1 + 1);
     if (snapInputLatLngToResult) {
       line1.add(point);
     }
-    var line2 = line.sublist(res.index + 1, line.length);
+    var line2 = line.sublist(res?.index ?? -1 + 1, line.length);
     if (snapInputLatLngToResult) {
       line2.insert(0, point);
     }
@@ -173,23 +173,23 @@ class VietmapPolyline {
 
   static num distanceRaw(LatLng from, LatLng to,
       [Unit unit = Unit.kilometers]) {
-    var dLat = degreesToRadians((to.latitude - from.latitude));
-    var dLon = degreesToRadians((to.longitude - from.longitude));
-    var lat1 = degreesToRadians(from.latitude);
-    var lat2 = degreesToRadians(to.latitude);
+    var dLat = _degreesToRadians((to.latitude - from.latitude));
+    var dLon = _degreesToRadians((to.longitude - from.longitude));
+    var lat1 = _degreesToRadians(from.latitude);
+    var lat2 = _degreesToRadians(to.latitude);
 
     num a =
         pow(sin(dLat / 2), 2) + pow(sin(dLon / 2), 2) * cos(lat1) * cos(lat2);
 
-    return radiansToLength(2 * atan2(sqrt(a), sqrt(1 - a)), unit);
+    return _radiansToLength(2 * atan2(sqrt(a), sqrt(1 - a)), unit);
   }
 
-  static num degreesToRadians(num degrees) {
+  static num _degreesToRadians(num degrees) {
     num radians = degrees.remainder(360);
     return radians * pi / 180;
   }
 
-  static num radiansToLength(num radians, [Unit unit = Unit.kilometers]) {
+  static num _radiansToLength(num radians, [Unit unit = Unit.kilometers]) {
     var factor = factors[unit];
     if (factor == null) {
       throw Exception("$unit units is invalid");
@@ -197,46 +197,46 @@ class VietmapPolyline {
     return radians * factor;
   }
 
-  static num bearingRaw(LatLng start, LatLng end, {bool calcFinal = false}) {
+  static num _bearingRaw(LatLng start, LatLng end, {bool calcFinal = false}) {
     // Reverse calculation
     if (calcFinal == true) {
-      return calculateFinalBearingRaw(start, end);
+      return _calculateFinalBearingRaw(start, end);
     }
 
-    num lng1 = degreesToRadians(start.longitude);
-    num lng2 = degreesToRadians(end.longitude);
-    num lat1 = degreesToRadians(start.latitude);
-    num lat2 = degreesToRadians(end.latitude);
+    num lng1 = _degreesToRadians(start.longitude);
+    num lng2 = _degreesToRadians(end.longitude);
+    num lat1 = _degreesToRadians(start.latitude);
+    num lat2 = _degreesToRadians(end.latitude);
     num a = sin(lng2 - lng1) * cos(lat2);
     num b = cos(lat1) * sin(lat2) - sin(lat1) * cos(lat2) * cos(lng2 - lng1);
 
-    return radiansToDegrees(atan2(a, b));
+    return _radiansToDegrees(atan2(a, b));
   }
 
-  static num bearing(LatLng start, LatLng end, {bool calcFinal = false}) =>
-      bearingRaw(start, end, calcFinal: calcFinal);
+  static num _bearing(LatLng start, LatLng end, {bool calcFinal = false}) =>
+      _bearingRaw(start, end, calcFinal: calcFinal);
 
-  static num calculateFinalBearingRaw(LatLng start, LatLng end) {
+  static num _calculateFinalBearingRaw(LatLng start, LatLng end) {
     // Swap start & end
-    num reverseBearing = bearingRaw(end, start) + 180;
+    num reverseBearing = _bearingRaw(end, start) + 180;
     return reverseBearing.remainder(360);
   }
 
   /// Calculates Final Bearing
-  static num calculateFinalBearing(LatLng start, LatLng end) =>
-      calculateFinalBearingRaw(start, end);
+  // static num _calculateFinalBearing(LatLng start, LatLng end) =>
+  //     _calculateFinalBearingRaw(start, end);
 
-  static num radiansToDegrees(num radians) {
+  static num _radiansToDegrees(num radians) {
     num degrees = radians.remainder(2 * pi);
     return degrees * 180 / pi;
   }
 
-  static LatLng destinationRaw(LatLng origin, num distance, num bearing,
+  static LatLng _destinationRaw(LatLng origin, num distance, num bearing,
       [Unit unit = Unit.kilometers]) {
-    num longitude1 = degreesToRadians(origin.longitude);
-    num latitude1 = degreesToRadians(origin.latitude);
-    num bearingRad = degreesToRadians(bearing);
-    num radians = lengthToRadians(distance, unit);
+    num longitude1 = _degreesToRadians(origin.longitude);
+    num latitude1 = _degreesToRadians(origin.latitude);
+    num bearingRad = _degreesToRadians(bearing);
+    num radians = _lengthToRadians(distance, unit);
 
     // Main
     num latitude2 = asin(sin(latitude1) * cos(radians) +
@@ -245,16 +245,16 @@ class VietmapPolyline {
         atan2(sin(bearingRad) * sin(radians) * cos(latitude1),
             cos(radians) - sin(latitude1) * sin(latitude2));
     return LatLng(
-      radiansToDegrees(latitude2).toDouble(),
-      radiansToDegrees(longitude2).toDouble(),
+      _radiansToDegrees(latitude2).toDouble(),
+      _radiansToDegrees(longitude2).toDouble(),
     );
   }
 
-  static LatLng destination(LatLng origin, num distance, num bearing,
+  static LatLng _destination(LatLng origin, num distance, num bearing,
           [Unit unit = Unit.kilometers]) =>
-      destinationRaw(origin, distance, bearing, unit);
+      _destinationRaw(origin, distance, bearing, unit);
 
-  static num lengthToRadians(num distance, [Unit unit = Unit.kilometers]) {
+  static num _lengthToRadians(num distance, [Unit unit = Unit.kilometers]) {
     num? factor = factors[unit];
     if (factor == null) {
       throw Exception("$unit units is invalid");
@@ -262,7 +262,7 @@ class VietmapPolyline {
     return distance / factor;
   }
 
-  static LatLng? intersects(List<LatLng> line1, List<LatLng> line2) {
+  static LatLng? _intersects(List<LatLng> line1, List<LatLng> line2) {
     if (line1.length != 2) {
       throw Exception('line1 must only contain 2 coordinates');
     }